+++ R/RcppExports.R
@@ -77,3 +77,22 @@ min_vl_dist <- function(s1, s2, elt_type = "pc", norm = "taxicab") {
     .Call('_minVL_min_vl_dist', PACKAGE = 'minVL', s1, s2, elt_type, norm)
 }
 
+#' Get minimal voice-leading set distances
+#'
+#' Equivalent to \code{\link{min_vl_dist()}} but vectorised.
+#' @param s1_list List of numeric vectors corresponding to the
+#' first pitch-class sets to be compared,
+#' with each number corresponding to either a pitch or a pitch class.
+#' Duplicates are permitted, and they will be retained.
+#' Order does not matter.
+#' @param s2 List of numeric vectors corresponding to the second
+#' pitch-class sets to be compared; see \code{s1}.
+#' @param elt_type Can be either \code{pitch} or \code{pc};
+#' determines whether \code{s1} and \code{s2} are interpreted as pitches or pitch classes.
+#' @param norm Can be either \code{euclidean}, \code{taxicab},
+#' or \code{infinity}. Each of these identify different norms.
+#' @export
+min_vl_dists <- function(s1_list, s2_list, elt_type = "pc", norm = "taxicab") {
+    .Call('_minVL_min_vl_dists', PACKAGE = 'minVL', s1_list, s2_list, elt_type, norm)
+}
+